The uploader is now a CsvImportUtil constructor argument
